Move swagger configs to a different class
